add warning, revert abort test
